0225-1

+++ client/src/admin/standBy.jsx
@@ -88,6 +88,8 @@ export default function Standby() {
 
                 <button onClick={(event)=>{
                   event.preventDefault()
+                  const ok = window.confirm("정말로 삭제하시겠습니까?")
+                  if (!ok) return
 
                   fetchApi('/auth/deletestandby', {
                     method: "DELETE",
